requirements.txt
langchain
langchain-openai
langchain-community
chromadb
pandas
python-dotenv
streamlit
matplotlib
